Fixed Markdown Guide in MenuHelper
