+++ app/i18n/_i18n.js
@@ -7,6 +7,12 @@ let config = new Config();
 let currentLangCode = config.get('lang') === undefined ? 'enUS' : config.get('lang');
 let lang = require(`./${getLangSource()}`);
 
+/**
+ * Gets the localization of a string.
+ * Uses the current language code in config.
+ * @param {string} key - The key of the localized string, in double-colon notation.
+ * @return The localization of the keyed string, or the most specific identifier of the key if no localization exists.
+ */
 function _(key) {
     // Re-cache lang if config lang has been updated
     if(config.get('lang') !== currentLangCode) {
@@ -31,6 +37,10 @@ function _(key) {
     }
 }
 
+/**
+ * Determines the source file of the current language.
+ * @returns The name of the JSON lang file associated with the lang code in config.
+ */
 function getLangSource() {
     let currentLang = langs.find((l) => {
         return l.id == currentLangCode;
